consolidate repo management in sampler.utils
