+++ mentat/sampler/utils.py
@@ -1,11 +1,53 @@
+import os
 from pathlib import Path
+from typing import Optional
 from uuid import uuid4
 
-from git import Repo  # type: ignore
+from git import GitCommandError, Repo  # type: ignore
 
 from mentat.errors import SampleError
 from mentat.git_handler import get_non_gitignored_files
 from mentat.utils import is_file_text_encoded
+
+CLONE_TO_DIR = Path(__file__).parent.parent.parent / "benchmark_repos"
+
+
+def clone_repo(
+    url: str, local_dir_name: str, refresh: bool = False, depth: int = 0
+) -> Path | None:
+    local_dir = CLONE_TO_DIR / local_dir_name
+    if os.path.exists(local_dir):
+        if refresh:
+            repo = Repo(local_dir)
+            repo.git.reset("--hard")
+            repo.git.clean("-fd")
+            repo.git.fetch("--all")
+    else:
+        if depth > 0:
+            repo = Repo.clone_from(url, local_dir, depth=depth)
+        else:
+            repo = Repo.clone_from(url, local_dir)
+    return local_dir
+
+
+def apply_diff_to_repo(diff: str, repo: Repo, commit: bool = False) -> str | None:
+    """Apply a git diff to a repo. If commit is True, commit the changes."""
+    temp_id = uuid4().hex
+    try:
+        # Save self.diff_merge_base to a temporary .diff file
+        with open(f".sample_{temp_id}.diff", "w") as f:
+            f.write(diff)
+        repo.git.execute(["git", "apply", f".sample_{temp_id}.diff"])
+        os.remove(f".sample_{temp_id}.diff")
+        if commit:
+            repo.git.add(".")
+            repo.git.commit("-m", f"sample_{temp_id}")
+    except GitCommandError as e:
+        try:
+            os.remove(f".sample_{temp_id}.diff")
+        except FileNotFoundError:
+            pass
+        return str(e)
 
 
 def get_active_snapshot_commit(repo: Repo) -> str | None:
