+++ tests/benchmarks/context_benchmark.py
@@ -14,8 +14,8 @@ from mentat.code_file_manager import CodeFileManager
 from mentat.config import Config
 from mentat.interval import Interval
 from mentat.llm_api import CostTracker, count_tokens, model_context_size, setup_api_key
+from mentat.sampler.utils import clone_repo
 from mentat.session_context import SESSION_CONTEXT, SessionContext
-from mentat.utils import clone_repo
 
 pytestmark = pytest.mark.benchmark
